A TodoMVC backend written in Go, running on GCP Managed VMs.
Implements the TodoMVC API (https://github.com/tastejs/todomvc-api) and
persists new Todos to the Datastore.

TodoMVC is linked as a git submodule, and most of its components are
served statically. In particular, we also do some monkeying with its
learn.json file, such that information relevant to our backend
implementation is displayed.